Fixed failing tests
On ruby 1.9.2p136, with RSpec 2.5.0.

There were initially 25 failing tests, based on two sets of problems.

The first of these were `@response.should have(1).key`, which I had to
append an "s" to. I don't like the plural, but it was the only way I got
the spec to pass. It gave an "ArgumentError: wrong number of arguments
(0 for 1)"

The second was related to procs/lambda, when no parameter was given in
the User model. E.g. ":if => lambda{ nil }" should be ":if => lambda{|u|
nil }", as it's called with a parameter every time.